feat(storage): 后端全链路测试通过
- 修复 SUM 类型转换（NUMERIC → BIGINT）
- 修复 STS 签名算法（自定义 RFC3986 编码）
- 测试脚本 6 步全部 PASS：订阅状态/兑换码/upload-token
- 新增 seed_subscription.sql 种子数据

+++ server/modules/app-subscription/src/repository.rs
@@ -71,7 +71,7 @@ impl SubRepo {
     /// 计算用户所有活跃订阅的总配额
     pub async fn sum_active_storage(db: &PgPool, user_id: i64) -> Result<i64, sqlx::Error> {
         let row: (i64,) = sqlx::query_as(
-            "SELECT COALESCE(SUM(p.storage_bytes), 0::BIGINT)
+            "SELECT COALESCE(SUM(p.storage_bytes)::BIGINT, 0::BIGINT)
              FROM user_subscriptions s
              JOIN subscription_plans p ON s.plan_id = p.id
              WHERE s.user_id = $1 AND s.status = 'active' AND s.expires_at > NOW()"
